backend/requirements.txt
# HandyWriterz Backend Dependencies

# Core Framework
fastapi>=0.104.1,<0.106.0
uvicorn[standard]>=0.24.0,<0.26.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0

# Redis
redis==5.0.1
aioredis==2.0.1

# AI/ML Libraries
openai>=1.86.0,<2.0.0
anthropic>=0.8.0,<1.0.0
google-generativeai>=0.3.0,<1.0.0

# Vector Storage
pgvector==0.2.3
chromadb==0.5.4

# File Processing
PyPDF2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6

# Cloud Storage
boto3==1.34.0
botocore==1.34.0

# Authentication & Security
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
cryptography>=36.0.0,<44.1

# HTTP Client
# let libraries choose a compatible httpx>=0.26,<1
aiohttp>=3.9.1,<4.0.0
requests>=2.31.0,<3.0.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
jsonschema==4.20.0
email-validator==2.1.0
Jinja2>=3.0.0,<4.0.0

# Monitoring & Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.39.0

# Background Tasks
celery==5.3.4
kombu==5.3.4

# Development Tools
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.0

# API Documentation
python-markdown==0.1.0

# Date/Time
python-dateutil==2.8.2
pytz==2023.3

# Telegram Integration
telethon==1.30.3
python-telegram-bot>=21.0.0,<22.0.0

# Math & Scientific
numpy==1.26.4
pandas==2.1.4
scipy==1.11.4

# Web Scraping (for research)
beautifulsoup4==4.12.2
scrapy>=2.11.0,<3.0.0

# Text Processing
nltk==3.8.1
spacy==3.7.2
textstat==0.7.3

# Async Support
asyncio==3.4.3
aiofiles==23.2.1

# Configuration
supabase==2.16.0

# LangChain extras
langchain-openai>=0.3.27,<0.4.0
langchain-anthropic>=0.3.17,<0.4.0
langchain-deepseek>=0.1.3,<0.2.0

# Other integrations
PyGithub==2.6.1
sentence-transformers>=5.0.0,<6.0.0
presidio-analyzer>=2.2.358,<3.0.0
presidio-anonymizer>=2.2.358,<3.0.0

mammoth
